Main Activities Page V.0.4
Implemented Upgrade Manager for each widget (Combinations does not have one, and it never will. It's broken enough as is). Implemented Combinations Task. Fixed Design Flaw in Page Update procedure. Removed Debug trackers due to In-Browser Axure having a "show console" in the top right corner which displays the relevant information for me.